[ISSUE 11] feat: Order Management BE

+++ src/main/java/com/ureca/myspring/repository/QuestionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ureca.myspring.dto.QuestionDTO;
 
-public interface QuestionRepository extends JpaRepository<QuestionDTO, Integer> {
-	List<QuestionDTO> findByStoreID(int storeID);
+public interface QuestionRepository extends JpaRepository<QuestionDTO, Long> {
+	List<QuestionDTO> findByStoreID(Long storeID);
 }
